bank/tx: makde IdempotenceKey routingNumbers long
they're long elsewhere

+++ bank-service/src/main/java/rs/banka4/bank_service/tx/data/IdempotenceKey.java
@@ -1,7 +1,7 @@
 package rs.banka4.bank_service.tx.data;
 
 public record IdempotenceKey(
-    int routingNumber,
+    long routingNumber,
     String locallyGeneratedKey
 ) {
 }
